Move web related util funcs to the web pkg
